pyproject.toml
[project]
name = "personal-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "markdown>=3.10.2",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "uuid-utils>=0.14.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.9",
]

[tool.ruff]
# Target version for Python features
target-version = "py311"
line-length = 88

[tool.ruff.lint]
# "E" and "F" are standard PEP8/Flake8 rules
# "I" handles import sorting (replaces isort)
# "UP" handles upgrade suggestions (modern Python syntax)
select = ["E", "F", "I", "UP", "N"]
ignore = ["D203"] # Example: ignoring specific docstring rules

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
